Moved all cloud related code to shared
