+++ src/operator/api/v1alpha1/intents_types.go
@@ -20,7 +20,7 @@ import (
 	"crypto/md5"
 	"encoding/hex"
 	"fmt"
-	kubernetes "github.com/otterize/intents-operator/src/operator/controllers/intents_reconcilers/otterizecloud/graphqlclient"
+	"github.com/otterize/intents-operator/src/shared/otterizecloud/graphqlclient"
 	"github.com/samber/lo"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
@@ -184,8 +184,8 @@ func (in *Intent) ResolveIntentNamespace(intentsObjNamespace string) string {
 	return intentsObjNamespace
 }
 
-func (in *ClientIntentsList) FormatAsOtterizeIntents() ([]kubernetes.IntentInput, error) {
-	otterizeIntents := make([]kubernetes.IntentInput, 0)
+func (in *ClientIntentsList) FormatAsOtterizeIntents() ([]graphqlclient.IntentInput, error) {
+	otterizeIntents := make([]graphqlclient.IntentInput, 0)
 	for _, clientIntents := range in.Items {
 		for _, intent := range clientIntents.GetCallsList() {
 			otterizeIntents = append(otterizeIntents, intent.ConvertToCloudFormat(clientIntents.GetServiceName()))
@@ -194,20 +194,20 @@ func (in *ClientIntentsList) FormatAsOtterizeIntents() ([]kubernetes.IntentInput
 	return otterizeIntents, nil
 }
 
-func (in *Intent) ConvertToCloudFormat(clientName string) kubernetes.IntentInput {
-	otterizeTopics := lo.Map(in.Topics, func(topic KafkaTopic, i int) kubernetes.KafkaConfigInput {
-		return kubernetes.KafkaConfigInput{
+func (in *Intent) ConvertToCloudFormat(clientName string) graphqlclient.IntentInput {
+	otterizeTopics := lo.Map(in.Topics, func(topic KafkaTopic, i int) graphqlclient.KafkaConfigInput {
+		return graphqlclient.KafkaConfigInput{
 			Name: topic.Name,
-			Operations: lo.Map(topic.Operations, func(op KafkaOperation, i int) kubernetes.KafkaOperation {
-				return kubernetes.KafkaOperation(op)
+			Operations: lo.Map(topic.Operations, func(op KafkaOperation, i int) graphqlclient.KafkaOperation {
+				return graphqlclient.KafkaOperation(op)
 			}),
 		}
 	})
-	intentInput := kubernetes.IntentInput{
+	intentInput := graphqlclient.IntentInput{
 		Client: clientName,
 		Server: in.Name,
-		Body: kubernetes.IntentBody{
-			Type: kubernetes.IntentType(in.Type),
+		Body: graphqlclient.IntentBody{
+			Type: graphqlclient.IntentType(in.Type),
 		},
 	}
 	if len(otterizeTopics) != 0 {
